Improved channel panel UI and added confirmation when closing
Closing a channel now prompts a confirmation dialog with some details.

The UI has also been reworked, the balance bar has been shrinked so that
it should not be mistaken as a separator element between channels. The
channel's balance and capacity are now more visible. The peer node id is
now more visible and is at the top of the channel's pane.

Also added the short channel id to the channel's pane.

fixes #690