Added node's aggregated balance in the status bar
Node's total balance is computed anew when the balance of a channel
changes.

Channel controller now contains a reference to the channel's actor and
also contains the balance/capacity of the channel. Thanks to this,
closing buttons actions can now defined within the controller, which
fixes an issue where the channel's details in the confirmation dialog
would not be up-to-date.

fixes #775