only support v3 hidden service
Eclair can connect to v2/v3, but will only open a v3 service. Instead of
trying our best to be compatible, we just require a minimal version for
the tor daemon.

The rationale for this change is that it allows for a significant
simplification, and is recommended in the Tor documentation:

> Since Tor 0.3.2 and Tor Browser 7.5.a5 56-character long v3 onion
addresses are supported and should be used instead.

source: https://www.torproject.org/docs/tor-onion-service.html.en#four