Actively clean the routing table in payment lifecycle
On android there are some cases where we are not aware that a local public
channel as in fact been closed. If this happens we will tell the router to
forget about it.

If the channel is in fact still alive, we will get it again via network
announcements anyway.