add a payment identifier
This is in preparation for a proper payments database.

There is no unique identifier for payments in LN protocol. Critically,
we can't use `payment_hash` as a unique id because there is no way to
ensure unicity at the protocol level.

Also, the generatl case for a "payment" is to be associated to multiple
`update_add_htlc`s, because of automated retries. We also routinely
retry payments, which means that the same `payment_hash` will be
conceptually linked to a list of lists of `update_add_htlc`s.

In order to address this, we introduce a payment id, which uniquely
identifies a payment, as-in a set of sequential `update_add_htlc`
managed by a single `PaymentLifecycle` that ends with a `PaymentSent` or
`PaymentFailed` outcome.

We can then query the api using either `payment_id` or `payment_hash`.
The former will return a single payment status, the latter will return a
set of payment statuses, each identified by their `payment_id`.

NB: This is just the first step, there is no payment db yet and the ids
are not actually used. Some db tests do not pass.

# Conflicts:
#	eclair-core/src/main/scala/fr/acinq/eclair/channel/ChannelTypes.scala
#	eclair-core/src/main/scala/fr/acinq/eclair/db/sqlite/SqliteAuditDb.scala
#	eclair-core/src/main/scala/fr/acinq/eclair/payment/PaymentEvents.scala
#	eclair-core/src/main/scala/fr/acinq/eclair/payment/PaymentLifecycle.scala
#	eclair-core/src/main/scala/fr/acinq/eclair/payment/Relayer.scala
#	eclair-core/src/test/scala/fr/acinq/eclair/api/JsonSerializersSpec.scala
#	eclair-core/src/test/scala/fr/acinq/eclair/channel/states/StateTestsHelperMethods.scala
#	eclair-core/src/test/scala/fr/acinq/eclair/channel/states/e/NormalStateSpec.scala
#	eclair-core/src/test/scala/fr/acinq/eclair/db/ChannelStateSpec.scala
#	eclair-core/src/test/scala/fr/acinq/eclair/db/SqliteAuditDbSpec.scala
#	eclair-core/src/test/scala/fr/acinq/eclair/payment/PaymentLifecycleSpec.scala
#	eclair-core/src/test/scala/fr/acinq/eclair/payment/RelayerSpec.scala
#	eclair-core/src/test/scala/fr/acinq/eclair/wire/ChannelCodecsSpec.scala