added a timeout for channel open request
Currently, if the peer is unresponsive (typically doesn't respond to
`open_channel` or `funding_created`), we wait indefinitely, or until the
connection closes.

This translates to an API timeout for users, and uncertainty about the
state of the channel.

This PR:
- adds an optional per-request timeout to the `open` endpoint, that will
actively cancel the channel opening if it takes too long before reaching
state `WAIT_FOR_FUNDING_CONFIRMED`.
- sets the akka http timeout (`akka.http.server.request-timeout`) to a
high value (10 min) under the assumption that we rely instead on the
underlying `ask` timeout.
- make the `ask` timeout configurable (`eclair.api.timeout`) with a
default to 60 seconds.