delay re-enabling of channel at reconnection
When switching from `SYNCING`->`NORMAL`, instead of emitting a new
`channel_update` with flag=enabled right away, we wait a little bit and
send it later.

This way, if a connection to a peer is unstable and we keep getting
disconnected/reconnected, we won't spam the network.

This extra delay allows us to remove the change made in #888, which was
a workaround in case we generated `channel_update` too quickly.

Also, increased refresh interval from 7 days to 10 days. There was no
need to be so conservative.