don't emit a `channel_update` at disconnection
Instead, wait for a payment to try to route through the channel and only
then reply with a disabled `channel_update` and broadcast it on the
network.

The reason in that in case of a disconnection, if noone cares about that
channel then there is no reason to tell everyone about its current
(disconnected) state.