re-emit `channel_update` at startup
We need to send those in order to properly initialize the `Router` and
the `Relayer` on startup. Otherwise they won't know about those
channels, and e.g. the `Relayer` will return `UnknownNextPeer` errors.

Note that this should have little or no impact to gossip because if
configuration hasn't changed, peer will already know the updates.

On the other hand, if some global parameters (like relaying fees) are
changed, it will cause the creation a new `channel_update` for all
channels.