Close the channel if the funding tx never confirms (#1034)
If we are fundee and after 5 days the funding isn't even in the mempool,
then we give up waiting and consider the channel closed. Note that if
the funding tx stays unconfirmed forever we won't give up waiting.

If we are funder, we never give up until the funding tx is double spent,
and we periodically republish it.

This applies to states `WAIT_FOR_FUNDING_CONFIRMED` and `CLOSING` (and
also `OFFLINE`/`SYNCING` when underlying state is
`WAIT_FOR_FUNDING_CONFIRMED`).

Also, added a generic way of passing context to `ElectrumClient`
requests/responses.

Fixes #1029.