Derive channel keys from funding pubkey
We now generate a random funding key for each new channel, and use its public key to deterministically derive
all channel keys and secrets. This will let us easily recover funds using DLP even if we've lost everything but
our seed: we just need to connect to the node we had a channel with, ask them to publish their commit tx, and
once we see it on the blockchain we can extract our funding pubkey, recompute channel keys and spend our output.