Added meta for native image generation
Added default parameters for SubstrateVM native-image ; these params will
automatically be used when generating the actual image, with a simple
command: `native-image -jar eclair-cli-x.y.z.jar`

We also use the assembly plugin to deliver a JAR containing all the dependencies.

Added a reflection json configuration file for native-image, describing
the classes used by Moshi for JSON parsing.

The project is also not a module of eclair anymore, because it slowed
down the build of this project (by adding several unused building
plugins) and added all the scalalib classes into the assembly JAR.