store channel data on restore
This may seem odd, but it is necessary in some cases when we have a data
format change that involves creating fields that contain a timestamp.

If we don't store that data, we will generate a new version of the data
at every restart, with a more recent timestamp.

It is the case for the `DATA_WAIT_FOR_FUNDING_CONFIRMED_COMPAT_01_Codec`
codec, or `DATA_CLOSING_COMPAT_06_Codec` codec, which results in
channels never been cleaned up.